refactor event config validations to separate functions
